+++ examples/database/dispatch_tasks.py
@@ -6,7 +6,6 @@ using the database module.
 import time
 
 import tensorflow as tf
-
 import tensorlayer as tl
 
 tl.logging.set_verbosity(tl.logging.DEBUG)
@@ -47,6 +46,5 @@ while db.check_unfinished_task(task_name='mnist'):
 
 # get the best model
 print("all tasks finished")
-sess = tf.InteractiveSession()
-net = db.find_top_model(sess=sess, model_name='mlp', sort=[("test_accuracy", -1)])
+net = db.find_top_model(model_name='mlp', sort=[("test_accuracy", -1)])
 print("the best accuracy {} is from model {}".format(net._test_accuracy, net._name))