+++ examples/spatial_transformer_network/tutorial_spatial_transformer_network_dynamic.py
@@ -1,7 +1,9 @@
 #! /usr/bin/python
 # -*- coding: utf8 -*-
 import time
+
 import numpy as np
+
 import tensorflow as tf
 import tensorlayer as tl
 from tensorlayer.layers import *
@@ -53,6 +55,7 @@ tl.vis.save_images(X_test_40[0:32], [4, 8], '_imgs_distorted.png')
 
 ##================== DEFINE MODEL ============================================##
 class Net(Model):
+
     def __init__(self):
         super(Net, self).__init__()