+++ examples/text_generation/data/__init__.py
@@ -1,4 +1,5 @@
 from __future__ import absolute_import
 
 from . import imagenet_classes
+
 # from . import