+++ examples/tutorial_work_with_onnx.py
@@ -105,7 +105,7 @@ In this tutorial, we continue our demonstration by performing inference using th
 >>>model = onnx.load('mnist.onnx')
 >>>tf_rep = prepare(model)
 >>>#Image Path
->>>img = np.load("./assets/image.npz")
+>>>img = np.load("./assets/image.npz", allow_pickle=True)
 >>>output = tf_rep.run(img.reshape([1, 784]))
 >>>print "The digit is classified as ", np.argmax(output)
 
@@ -117,13 +117,13 @@ You will get the information in your console:
 import time
 
 import numpy as np
-import tensorflow as tf
-from tensorflow.python.tools.freeze_graph import freeze_graph as _freeze_graph
 
 import onnx
+import tensorflow as tf
 import tensorlayer as tl
 from onnx_tf.backend import prepare
 from onnx_tf.frontend import tensorflow_graph_to_onnx_model
+from tensorflow.python.tools.freeze_graph import freeze_graph as _freeze_graph
 
 tf.logging.set_verbosity(tf.logging.DEBUG)
 tl.logging.set_verbosity(tl.logging.DEBUG)
@@ -317,7 +317,7 @@ def convert_onnx_to_model(onnx_input_path):
     model = onnx.load(onnx_input_path)
     tf_rep = prepare(model)
     # Image Path
-    img = np.load("./assets/image.npz")
+    img = np.load("./assets/image.npz", allow_pickle=True)
     output = tf_rep.run(img.reshape([1, 784]))
     print("The digit is classified as ", np.argmax(output))