+++ tensorlayer/layers/scale.py
@@ -2,7 +2,6 @@
 # -*- coding: utf-8 -*-
 
 import tensorflow as tf
-
 from tensorlayer import logging
 from tensorlayer.initializers import constant
 from tensorlayer.layers.core import Layer
@@ -22,7 +21,7 @@ class Scale(Layer):
     name : a str
         A unique layer name.
 
-    Examples:
+    Examples
     ----------
     >>> inputs = tl.layers.Input([8, 3])
     >>> dense = tl.layers.Dense(n_units=10)(inputs)