+++ tensorlayer/models/core.py
@@ -3,9 +3,8 @@ from abc import abstractmethod
 from queue import Queue
 
 import tensorflow as tf
-from tensorflow.python.framework import ops as tf_ops
-
 import tensorlayer as tl
+from tensorflow.python.framework import ops as tf_ops
 from tensorlayer import logging
 from tensorlayer.files import utils
 from tensorlayer.layers import Layer, ModelLayer
@@ -402,7 +401,7 @@ class Model(object):
                 if layer.trainable_weights is not None:
                     self._trainable_weights.extend(layer.trainable_weights)
 
-        return self._trainable_weights
+        return self._trainable_weights.copy()
 
     @property
     def nontrainable_weights(self):
@@ -416,7 +415,7 @@ class Model(object):
                 if layer.nontrainable_weights is not None:
                     self._nontrainable_weights.extend(layer.nontrainable_weights)
 
-        return self._nontrainable_weights
+        return self._nontrainable_weights.copy()
 
     @property
     def all_weights(self):
@@ -430,7 +429,7 @@ class Model(object):
                 if layer.all_weights is not None:
                     self._all_weights.extend(layer.all_weights)
 
-        return self._all_weights
+        return self._all_weights.copy()
 
     @property
     def config(self):
@@ -443,13 +442,14 @@ class Model(object):
                 _config.update({"name": None})
             else:
                 _config.update({"name": self.name})
-            # versionInfo = {
-            #     "tensorlayer_version": tl.__version__,
-            #     "backend": "tensorflow",
-            #     "backend_version": tf.__version__,
-            #     "training_device": "gpu",
-            # }
-            # _config.update(versionInfo)
+            version_info = {
+                "tensorlayer_version": tl.__version__,
+                "backend": "tensorflow",
+                "backend_version": tf.__version__,
+                "training_device": "gpu",
+                "save_date": None,
+            }
+            _config["version_info"] = version_info
             # if self.outputs is None:
             #     raise RuntimeError(
             #         "Dynamic mode does not support config yet."
@@ -591,6 +591,15 @@ class Model(object):
             layer._fix_nodes_for_layers()
         self._nodes_fixed = True
 
+    def __setattr__(self, key, value):
+        if isinstance(value, Layer):
+            if value._built is False:
+                raise AttributeError(
+                    "The registered layer `{}` should be built in advance. "
+                    "Do you forget to pass the keyword argument 'in_channels'? ".format(value.name)
+                )
+        super().__setattr__(key, value)
+
     def __repr__(self):
         # tmpstr = self.__class__.__name__ + '(\n'
         tmpstr = self.name + '(\n'
@@ -669,6 +678,8 @@ class Model(object):
 
         visited_node_names = set()
         for out_node in output_nodes:
+            if out_node.visited:
+                continue
             queue_node.put(out_node)
 
             while not queue_node.empty():